+++ src/components/LibraryItem.js
@@ -4,8 +4,14 @@ import CloseIcon from "../components/icons/CloseIcon";
 import { useDispatch } from "react-redux";
 import { libraryActions } from "../store/library-slice";
 import { suppliersActions } from "../store/suppliers-slice";
+import swal from "sweetalert";
 
 const LibraryItem = (props) => {
+  const materialName = props.name;
+  const materialSupplier = props.supplier;
+
+  console.log(`You are about to delete ${materialName} by ${materialSupplier}`);
+
   const dispatch = useDispatch();
 
   const targetEditHandler = (ev) => {
@@ -23,25 +29,38 @@ const LibraryItem = (props) => {
     const targetStart = ev.target.closest("div[class*='LibraryItem_item']")
       .dataset.order;
 
-    dispatch(
-      libraryActions.deleteMaterial({
-        index: !props.supEdit ? targetStart : "",
-        supplier: props.supplier,
-        name: props.name,
-        collection: props.collection,
-        category: props.category,
-      })
-    );
-
-    dispatch(
-      suppliersActions.deleteMaterial({
-        index: props.supEdit ? targetStart : "",
-        name: props.name,
-        supplier: props.supplier,
-        collection: props.collection,
-      })
-    );
-    //}
+    swal({
+      title: `You are about to delete ${materialName} by ${materialSupplier}!`,
+      text: "Once deleted, material informations will be lost!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        swal(`${materialName} was deleted!`, {
+          icon: "success",
+        });
+        dispatch(
+          libraryActions.deleteMaterial({
+            index: !props.supEdit ? targetStart : "",
+            supplier: props.supplier,
+            name: props.name,
+            collection: props.collection,
+            category: props.category,
+          })
+        );
+        dispatch(
+          suppliersActions.deleteMaterial({
+            index: props.supEdit ? targetStart : "",
+            name: props.name,
+            supplier: props.supplier,
+            collection: props.collection,
+          })
+        );
+      } else {
+        swal("Your material informations are safe!");
+      }
+    });
   };
   return (
     <div className={classes.item} data-order={props.dataset}>
@@ -58,7 +77,11 @@ const LibraryItem = (props) => {
       <div className={classes.materialSupplier}>{props.supplier}</div>
       <div className={classes.materialCertificate}>{props.certificate}</div>
       <div className={classes.materialInfo}>{props.info}</div>
-      <img className={classes.materialImage} src={props.imageUrl} />
+      <img
+        className={classes.materialImage}
+        alt={`Material name: ${materialName}. Supplier name: ${materialSupplier}`}
+        src={props.imageUrl}
+      />
       <a className={classes.materialLink} href={props.link} target="_blank">
         Link
       </a>
